Sho03

+++ resources/views/posts/index.blade.php
@@ -1,36 +1,33 @@
-<!DOCTYPE html>
-<html lang="{{ str_replace('_', '-', app()->getLocale()) }}">
-    <head>
-        <meta charset="utf-8">
-        <title>Blog</title>
-        <!-- Fonts -->
-        <link href="https://fonts.googleapis.com/css?family=Nunito:200,600" rel="stylesheet">
-    </head>
-    <body>
-        <h1>レバテックチーム開発</h1>
-        <h2>投稿一覧ページ</h2>
-        <div class='posts'>
-            @foreach($posts as $post)
-                <div class='post'>
-                    <h2 class='title'>
-                        タイトル：<a href="/posts/{{ $post->id }}">{{ $post->title }}</a>
-                    </h2>
-                    <p class='body'>本文：{{ $post->body}}</p>
+@extends('layouts.app')　　　　　　　　　　　　　　　　　　・・・　２
+
+@section('content')
+    <h1>レバテックチーム開発</h1>
+    <h2>投稿一覧ページ</h2>
+    <div class='posts'>
+        @foreach($posts as $post)
+            <div class='post'>
+                <h2 class='title'>
+                    タイトル：<a href="/posts/{{ $post->id }}">{{ $post->title }}</a>
+                </h2>
+                <p class='body'>本文：{{ $post->body}}</p>
+                @if (Auth::check() && $user->host_flg == '1')
                     <form action="/posts/{{ $post->id }}" id="form_{{ $post->id }}"  method="post" style="display:inline">
                         @csrf
                         @method('DELETE')
                         <button type="button" onClick="deletePost({{$post->id}});">削除</button> {{--script内に定義したdeletePostを使用している--}}
                     </form>
-                </div>
-            @endforeach
-        </div>
-        <div class='paginate'>
-            {{ $posts->links() }}
-        </div>
+                @endif
+            </div>
+        @endforeach
+    </div>
+    <div class='paginate'>
+        {{ $posts->links() }}
+    </div>
+    @if (Auth::check() && $user->host_flg == '1')
         <div>
             [<a href='/posts/create'>新規作成</a>]
         </div>
-    </body>
+    @endif
     <script>
         function deletePost(post_id) {
             form = document.getElementById('form_' + post_id);  //各投稿ごとのdeleteのformを取得
@@ -41,4 +38,4 @@
             }
         }
     </script>
-</html>
+@endsection
